requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
autoflake==2.3.1
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.20.0
azure-mgmt-core==1.5.0
azure-storage-blob==12.19.0
azureml-mlflow==1.59.0.post1
beautifulsoup4==4.13.3
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cmudict==1.0.32
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
databricks-sdk==0.44.1
dataclasses-json==0.6.7
datasets==3.3.2
debugpy==1.8.12
decorator==5.2.1
Deprecated==1.2.18
dill==0.3.8
distlib==0.3.9
docker==7.1.0
evaluate==0.4.3
executing==2.2.0
fastapi==0.115.11
filelock==3.17.0
filetype==1.2.0
flake8==7.1.2
flake8-bugbear==24.12.12
Flask==3.1.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
identify==2.6.8
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isodate==0.7.2
isort==6.0.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpickle==4.0.2
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.6
langsmith==0.3.11
Mako==1.3.9
Markdown==3.7
markdownify==1.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mlflow==2.20.3
mlflow-skinny==2.20.3
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.1.0
multiprocess==0.70.16
mypy==1.15.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
numpy==2.2.3
oauthlib==3.2.2
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
pre_commit==4.1.0
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.28.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pyphen==0.17.2
pytest==8.3.4
pytest-docker==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rouge_score==0.1.2
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.12
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.0
sympy==1.13.1
tenacity==9.0.0
textstat==0.7.5
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.2
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
xxhash==3.5.0
yamlfmt==1.1.1
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
